Initial Commit
mqtt_script_runner.py runs an MQTT client which connects to
io.adafruit.com (but this can easily be replace with another service),
the authentication details need to be stored in <working_dir>/auth
like so:
<username>\n
<password>\n

The authentication details are the adafruit username and AIO key

The client subscribes to the topics listed in
<working_dir>/topics
And can run scripts which are in the <working_dir>/<scripts>/
directory.

When the client receives new data it finds the closest match
between the new data and the list of script names with a certain
tolerance so "bake potato" doesn't open your garage or turn off
your computer.

Signed-off-by: Amit Matityahu <amit2129@gmail.com>